backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.7.1
python-multipart==0.0.9
sqlalchemy==2.0.30
pymysql==1.1.1
python-dotenv==1.0.1
certifi>=2024.2.2
